ClubMoney/requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
Django==4.1.3
django-cors-headers==3.13.0
django-cron==0.6.0
django-crontab==0.7.1
django-fcm==0.1.1
django-push-notifications==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangotoolbox==1.8.0
fcm-django==1.0.12
firebase-admin==5.4.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.21.0
idna==3.4
kombu==5.2.4
msgpack==1.0.4
Pillow==9.4.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.0
pusher-push-notifications==2.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.2
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6